🏕️ Rename feedParser -> feedFetcher
The name is more fitting.
