+++ src/Feed/Application/Command/Feed/Handler/FetchFeedHandler.php
@@ -6,7 +6,7 @@ use App\Common\Infrastructure\Messenger\CommandBus\AsCommandHandler;
 use App\Common\Infrastructure\Messenger\EventBus\EventBus;
 use App\Feed\Application\Command\Feed\FetchFeedCommand;
 use App\Feed\Application\Event\Feed\FeedItemWasFetchedEvent;
-use App\Feed\Application\FeedParser\FeedParser;
+use App\Feed\Application\Service\FeedFetcher\FeedFetcher;
 use App\Feed\Domain\Source\SourceId;
 use App\Feed\Domain\Source\SourceRepository;
 
@@ -14,7 +14,7 @@ final readonly class FetchFeedHandler
 {
     public function __construct(
         private SourceRepository $sourceRepository,
-        private FeedParser $feedParser,
+        private FeedFetcher $feedFetcher,
         private EventBus $eventBus,
     ) {
     }
@@ -24,7 +24,7 @@ final readonly class FetchFeedHandler
     {
         $source = $this->sourceRepository->findOrThrow(new SourceId($command->sourceId));
 
-        $feedItems = $this->feedParser->fetchFeed($source->getName(), $source->getUrl());
+        $feedItems = $this->feedFetcher->__invoke($source->getName(), $source->getUrl());
 
         foreach ($feedItems as $feedItem) {
             $this->eventBus->dispatch(
