--- src/Feed/Application/FeedParser/FeedParser.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Feed\Application\FeedParser;
-
-interface FeedParser
-{
-    /**
-     * @return List<FeedItem>
-     */
-    public function fetchFeed(string $source, string $url): array;
-}
