+++ src/Feed/Application/Service/FeedFetcher/FeedFetcher.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Feed\Application\Service\FeedFetcher;
+
+interface FeedFetcher
+{
+    /**
+     * @return List<FeedItem>
+     */
+    public function __invoke(string $source, string $url): array;
+}
