+++ src/Feed/Application/Service/FeedItemNormalizer/FeedItemNormalizer.php
@@ -2,7 +2,7 @@
 
 namespace App\Feed\Application\Service\FeedItemNormalizer;
 
-use App\Feed\Application\FeedParser\FeedItem;
+use App\Feed\Application\Service\FeedFetcher\FeedItem;
 
 interface FeedItemNormalizer
 {
