+++ tests/Feed/Application/Service/FeedItemNormalizer/ShortenSummaryNormalizerTest.php
@@ -2,7 +2,7 @@
 
 namespace Feed\Application\Service\FeedItemNormalizer;
 
-use App\Feed\Application\FeedParser\FeedItem;
+use App\Feed\Application\Service\FeedFetcher\FeedItem;
 use App\Feed\Application\Service\FeedItemNormalizer\ShortenSummaryNormalizer;
 use PHPUnit\Framework\TestCase;
